feat: Don't instrument generated code
Use the fact that generated code won't map to a source file (if
a source map is present) to automatically ignore it.

Also, instead of crawling the entire source tree and fixing up all
node locations with the source map, only consult it when necessary
for accurate location of recorded functions.

Resolving locations with source maps is expensive and most of the
work involved with the tree crawl has been wasted.