fix: Support Jest version 27
In older versions of Jest circus is run in a test environment
which at this point doesn't have access to AppMap globals.
This change passes AppMap globals into jest environments.

Note I decided not to add a test for Jest 27 to avoid dependency
bloat; seeing as Jest 27 is unmaintained and the last release
was over a year ago it doesn't seem worth the trouble.

Fixes #53